Separate query resolvers by domain models
