+++ vehicle-depot-ms/src/main/java/com/plutoz/carryit/vehicledepot/graphql/query/VehicleQueryResolver.java
@@ -1,11 +1,9 @@
 package com.plutoz.carryit.vehicledepot.graphql.query;
 
 import com.coxautodev.graphql.tools.GraphQLQueryResolver;
-import com.plutoz.carryit.vehicledepot.domain.Depot;
-import com.plutoz.carryit.vehicledepot.graphql.query.pagination.PagedResponse;
-import com.plutoz.carryit.vehicledepot.graphql.query.pagination.Page;
 import com.plutoz.carryit.vehicledepot.domain.Vehicle;
-import com.plutoz.carryit.vehicledepot.service.DepotService;
+import com.plutoz.carryit.vehicledepot.graphql.query.pagination.Page;
+import com.plutoz.carryit.vehicledepot.graphql.query.pagination.PagedResponse;
 import com.plutoz.carryit.vehicledepot.service.VehicleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -13,20 +11,13 @@ import org.springframework.stereotype.Component;
 import java.util.Optional;
 
 @Component
-public class QueryResolver implements GraphQLQueryResolver {
+public class VehicleQueryResolver implements GraphQLQueryResolver {
 
-    private final DepotService depotService;
     private final VehicleService vehicleService;
 
     @Autowired
-    public QueryResolver(DepotService depotService, VehicleService vehicleService) {
-        this.depotService = depotService;
+    public VehicleQueryResolver(VehicleService vehicleService) {
         this.vehicleService = vehicleService;
-    }
-
-    PagedResponse<Depot> getAllDepot(Optional<Page> page) {
-        var pagedResult = depotService.findAll(page.orElse(Page.defaultPage()).asPageable());
-        return PagedResponse.from(pagedResult);
     }
 
     PagedResponse<Vehicle> getAllVehicle(Optional<Page> page) {
